Pulled out another test (still hard-coded) and switched to DCT-2/3.
The cos_test is meant to test that the poisson kernel is being
applied to simple functions (e.g. cos(x)) properly.  It works.

In the process of using cos_test, I found a bug in the z-direction
(walls) transform: since the element centers are offset from the
walls, using DCT-2/3 makes a lot more sense than DCT-1.