Moved spectral_solve into hsmg_coarse solve as a preconditioner for gmres.
This is probably where it belonged all along: operating on the coarse
mesh produced by the hybrid Schwarz multi-grid in h1mg_solve.

The coarse mesh has a 2x2x2 basis in each element, so there is some
dssum hacking to get a single point, corresponding to the root, for
each element.  The result is a mesh that is missing one face, so the
boundary conditions need to be fixed.