Switching to subspace diagonalization of approx.
The contents of approx are very similar.  The previous scheme
did a good job subtracting off the most recent solution, but the
residual was mostly orthogonal to the rest of approx and there
was little return after that.

The Helmholtz operator is now diagonalized in the space spanned
by approx(:,:,1), which yeilds orthogonal eigenvectors that don't
suffer this problem quite as badly.  However, the high overlap
in approx can cause some precision issues, as the subspace
Helmholtz operator has very little variation in its elements.

This commit is highly un-optimized; next commit will be faster.