The contents of the approximation space cycle.
Since the projectors are independent of one another, we can just
replace the oldest one with a new one, keeping the whole space
filled.  This avoids jumps in the iteration count.

This is implemented through a new approx_space type which holds
the projectors, an index array to avoid large copies, and the
subspace operator.