Ported direct subspace orthogonalization from NekBox.
The operator is computed in the subspace of the previous solutions
and directly diagonalized.  Then, the residual is projected onto
the eigenvectors of the subspace operator, the approximate solution
is formed, and the contribution to the residual is removed.

This seems to be giving the right answer, but isn't quite as good
as the original.  I'm not sure if this is a real bug, or subtle
numerical differences because the time-step is so small.  More
validation to come.