Added new "series" input which is used to name fld files (not rea).
This is mostly to help with compatability with swift, which requires
each job to have a unique rea file.  Previously, with rea and
fld files named with the same prefix, different jobs were producing
outputs with different prefixes, which required copies for restarts
etc, etc...

If series is not specified (as the third line of the SESSION file)
then it defaults to the session and the old behavior is maintained.