Finally working; changing default PP kind to 4.
This commit includes a single precision version of mxm, replaces
some custom transposes with intrinsics, and a few bugfixes. At this
point, PP /= 8 actually works.

For the iturb case, the result validates and takes the same number
of pressure iterations, indicating that SP is sufficient.  Part
of this commit is to see what travis ends up saying.